dd is consistently used with byte seek
The old style was to play with "bs" to achieve the proper seek goal.
This also then influenced copy performance. By specifically requesting
byte seeks and skips we don't have to mutilate "bs" and get a more
consistent results. To ensure math is done properly for the byte
offsets in all cases we use "bc" instead of expr. expr has a rather
peculiar way of handling return codes and the problem can be eliminated
this way in a very deterministic fashion.

Signed-off-by: Heinz Wrobel <Heinz.Wrobel@nxp.com>
Signed-off-by: Alexandru-Catalin Ionita <alexandru-catalin.ionita@nxp.com>
Issue: ALB-10781