converted expr math into bc math
This is a fairly straight conversion and eliminates the obscure
"|| true" hack we had to add in earlier. Even though bc at first is
a tiny bit more overhead in calculation, it makes the math much more
readable. This is good for maintenance and permits a proper review of
what is actually happining.

Signed-off-by: Heinz Wrobel <Heinz.Wrobel@nxp.com>
Issue: ALB-10781