Use single function to issue all dd invocations
The problem with individual dd is that a) they don't really show up in
the log in a readable form. So by using a general function that issues
a bbnote and handles things we get a proper log.
The common dd options are also then in one place and there is less
potential for copy/paste errors.

Signed-off-by: Heinz Wrobel <Heinz.Wrobel@nxp.com>
Signed-off-by: Alexandru-Catalin Ionita <alexandru-catalin.ionita@nxp.com>
Issue: ALB-10781