Numerous quoting cleanups and one dd replaced
While the files and dirs used in practice do not have spaces, there were
numerous quoting problems that could have led to weird results if
spaces would suddenly be present.
This commit is mainly a cleanup to ensure that all scripting is robust.

An invisible functional change is that backtick quoting also has been
changed to $() for more robustness in maintenance. This, together with
quoting, however exhibits a weakness in bitbake parsing of complex
shell expressions. Note the implementation change in
update_rootfs_name(). Rather than returning by printing the string, we
now return the variable because we can't quote $() or bitbake optimizes
away the function we call in $().

One user visible functional change in this commit is that a left over
dd call is now also using generate_sdcardimage_entry_raw(), which
ensures that now all image building elements should show up in the log
with their files and offsets.

Signed-off-by: Heinz Wrobel <Heinz.Wrobel@nxp.com>
Signed-off-by: Alexandru-Catalin Ionita <alexandru-catalin.ionita@nxp.com>
Issue: ALB-10781