dm-verity: roothash signing
Currently, for a secure boot, the chain of trust is broken on
rootfs. A bad actor can replace the rootfs, metadata file system and
root hash input.

Avoid this by signing the root hash and verifying it before
mounting the rootfs.

To activate the roothash signing process, add "root-hash-sign" and
"dm-verity" to DISTRO_FEATURES.

Issue: ALB-10668
Signed-off-by: Alexandru-Catalin Ionita <alexandru-catalin.ionita@nxp.com>