Unified raw dd disk access for flash and SD card images
As it turns out, both the flash image class and the SD card image class
did similar to mostly identical things when it comes to writing raw
information with dd to the image. The problem however was that they did
things "sort of" the same, but not quite. Problems with that were, e.g.,
- flashimage did overlap checking, SD didn't
- there was a silent fallback to DEPLOY_DIR_IMAGE for files rather than
  doing an unambiguous file selection. Risky
- Some dd showed up with a readable log, some didn't

This commit factors out a common set of functions to do dd insertion of
data into images with overlap checking.
It then modifies the flashimage and sdcard classes to make use of those
functions.

The net result is that all dd's that write image date are now properly
logged and checked for both flash image and sdcard classes.
The class implementations should also be cleaner to maintain now.

Signed-off-by: Heinz Wrobel <Heinz.Wrobel@nxp.com>
Issue: ALB-10781