lx2160bluebox3: Fix enumeration of mmcblk under Linux Qoriq
No aliases were defined that cleanly mapped the two SDHC ports of the
SoC. Given the undefined nature of how the dtb is processed by the
Linux Kernel this lead to mmcblk0 being suddenly the second interface
on the SoC, breaking scripting and booting.

With aliases 'mmc0' and 'mmc1' defined in the dtb we should now have
a proper mapping in all cases.

Issue: ALB-10780
Signed-off-by: Heinz Wrobel <Heinz.Wrobel@nxp.com>