sdcard class dependencies were missing and not robust
First, the code is placing a kernel but never actually states the
dependency on the kernel recipe. Net result is that on clean builds
you can run into a race condition where the kernel build is not yet
done, but the SD card image is already generated ... leading to failure.
Specifying a dependency on the chosen kernel fixes that race condition.

Second, the class relied on host tools without even clearly stating it
when there is no need to do so.
We now specify the dependencies for the native version of mtools,
parted, and dosfstools to avoid relying on unknown host Linux configs.
This makes things much more robust.

Issue: ALB-10800
Signed-off-by: Heinz Wrobel <Heinz.Wrobel@nxp.com>