meta-alb: remove Eth name rewrite in complex images on QorIQ targets
Given the prior commits, we now have a fairly stable setup of using
udev-extraconf consistently. This means that we can now consistently
make sure that udev-rules-qoriq, which does the Ethernet name rewrite
is not applied anymore for non core-image* images.

This is the first step to full consistency as we also need to deal
with image recipes.

Issue: ALB-10792
Signed-off-by: Heinz Wrobel <Heinz.Wrobel@nxp.com>