linux-qoriq: Updated the Aquantia driver to support rate adaption more generally
The original AQ driver would lock the line rate to the SerDes rate.
In a prior patch set, I had released that lock for the 405 only to
support the LS2 RDB designs and derivatives. The problem was that the
same topic really affects the >=Gen2 parts, too, e.g., AQR107.
This patch now integrates the support and will no longer lock the line
rate if the rate adapation is enabled on >=Gen2 parts. It's still a bit
awkward, but at least a somewhat straightforward and deterministic
solution.

Issue: ALB-10820
Signed-off-by: Heinz Wrobel <Heinz.Wrobel@nxp.com>