u-boot: Fixed NOR flash config for BB2
The BB2 NOR flash configuration was broken and U-Boot couldn't be used
to update the NOR flash. So I had to update the ls2084abbmini config
appropriately and cleaned the BB support patches in the Yocto tree.

The number of patches increased because I intentionally split the
U-Boot patches into the new files added per board and the config
changes in shared files to enable the board. This should permit
easier incremental updates of board specific elements without having
to touch the shared U-Boot config or vice versa. On rev'ing U-boot
this makes for simpler and less risky Yocto updates.

Issue: ALB-10840
Signed-off-by: Heinz Wrobel <Heinz.Wrobel@nxp.com>