qoriq-atf: LS2 boot source detection is broken
On the LS2 family of SoCs, the boot source register is read with a dcfg
related function before the dcfg subsystem has been initialized. This
resulted in a nonsense boot source be read through an uninitialized
pointer.

The solution is to move the call down a bit, equivalent to what is done
on LX2.

Issue: ALB-10847
Signed-off-by: Heinz Wrobel <Heinz.Wrobel@nxp.com>
Signed-off-by: Florin Buica <Florin.Buica@nxp.com>