image_types_fsl_sdcard: silent exit if OFFSET variable is not set
The sdcard image with PFE enabled could not be created because the
SDCARDIMAGE_EXTRA_OFFSET{1-9} variables were not set. This is because
the SDCARDIMAGE_EXTRA_OFFSET variable is assigned by
FLASHIMAGE_EXTRA_OFFSET{1-9} variable which is not set when
building sdcard. Previously it was working because the check for the
offset was not done correctly.

On the SDcard image, firmwares are placed on the boot partition and not
inside the image. Thus, when  the image is built, the OFFSET variable
might be empty. If it's empty, don't break the build. Instead, silent
exit.

Issue: ALB-10781
Signed-off-by: Alexandru-Catalin Ionita <alexandru-catalin.ionita@nxp.com>