security: optee: make tee-supplicant a single service
The latest changes in the meta-arm layer saw the convertion of the
tee-supplicant single service unit file into a template for muliple
service instances (hence the '@' adfition in the name). The upstream
commit '1bff41ee optee-client: start tee-supplicant.service when
teeprivX dev is detected' introduced the change and requires having a
kernel udev rule which would be activated upon the detection of
a teeprivX device. However, the udev rule is integrated in the recipes
for trusted services of the same layer.

This commits reverts back to the tee-supplicant being a single service
because:
1. The meta-alb project is configured to have a single tee provider at
the moment: OP-TEE. Having multiple instances of tee-supplicant is an
unnecesarry addition.
2. Bringing the trusted-services recipes into meta-alb seems unnecesarry
for just a single udev rule file.

Issue: ALB-10784
Signed-off-by: Bogdan Roman <bogdan-gabriel.roman@nxp.com>