perf: make pkg-config available for kernels > 6.4
With kernels < 6.4 pkg-config is only used for the gtk+-2.0 integration
which is disabled in the OE config. Thus the issue wasn't exposed so
far. With kernel 6.4 pkg-config is additionally used for libtraceevent
and libtracefs.

This commit is a backport from latest poky Yocto layer:
(commit id: '03adf4b267b3')

Issue: ALB-10898
Signed-off-by: Ciprian Marian Costea <ciprianmarian.costea@nxp.com>