Clearly define kernel version for BlueBox LS targets
Issue: ALB-11023

Previously there was no preferred version for the BB LS side machines.
This worked because S32 and LS side were using the same base version.
Given different update cycles this cannot always be guaranteed, so
we specify clearly which kernel version we need.

Given that we use both linux-qoriq and linux-flash it is important to
ensure that both refer to the same version.

Signed-off-by: Heinz Wrobel <Heinz.Wrobel@nxp.com>