Fix rootfs creation when secboot is enabled in DISTRO_FEATURES
Issue: ALB-11060

arm-trusted-firmware FILES:${PN} needs to be initialized with
empty content in order to allow different content be split on
different other packages-split , otherwise would contain most
of the files installed while the rest of other packages-split
would be empty. Feature secboot needs to install the public key
/etc/keys/secboot/fip_rsa2048_public.pem  contained by package
arm-trusted-firmware-secboot in rootfileystem in order to be
used later for enabling secboot.

Signed-off-by: Mihai Adrian Enache <mihai-adrian.enache@nxp.com>