fix kernel install when folder modules/${KERNEL_VERSION}/kernel does not exist
Issue: ALB-11098

The kernel.bbclass from current yocto version expects folder
modules/${KERNEL_VERSION}/kernel created in order to remove it.
If it does not exist, it will fail. This is problematic since
one kernel version leaves that folder created while the other
cleans after itself. Fix is to remove the folder only if it exists.

Signed-off-by: Mihai Adrian Enache <mihai-adrian.enache@nxp.com>