recipes-devtools: qemu: Add DEPENDS on xen-tools for class-target
The libs and public headers exported by xen-tools after compilation
need to be used when compiling QEMU to ensure ABI compatibility,
for the QEMU which will be deployed & run on target.

Add a DEPENDS on "xen-tools" to make sure they are provided
in QEMU's sysroot if we're building for target.

Issue: ALB-11340
Signed-off-by: Andrei Cherechesu <andrei.cherechesu@nxp.com>