meta-alb-bsp: fix tf-a dependency on u-boot's sysroot for hse fw
The u-boot recipe creates a configuration file, u-boot-s32-*.cfgout
which is later used by tf-a during the compilation step. When hse
distro feature is used, the path to the hse firmware is included in this
configuration file. However, the path is relative to the u-boot's
sysroot: this leads to synchronization problems when tf-a tries to
access u-boot's sysroot.
Fix this issue by adding a dependency to hse-firmware in the tf-a recipe
so that tf-a can update the path in the configuration file with
the location of the firmware relative to its sysroot.

Issue: ALB-11397
Signed-off-by: Bogdan Roman <bogdan-gabriel.roman@nxp.com>